+++ .env
@@ -21,7 +21,6 @@ NEXT_PUBLIC_GA_MEASUREMENT_ID="G-XXXXXXXXXX"
 # Database
 DATABASE_URL="postgresql://postgres@localhost:5432/startkit_development"
 # DATABASE_URL_POOLED=
-# DATABASE_URL_EDGE=
 
 # Authentication JWT Secret (`openssl rand -base64 32`)
 AUTH_SECRET=
