+++ app/(legal)/layout.tsx
@@ -5,12 +5,6 @@ import { Footer } from "@/components/layout/footer"
 import { Button } from "@/components/ui/button"
 
 import type { ReactNode } from "react"
-
-/**
- * Use the edge runtime
- */
-export const runtime = "edge"
-
 interface Props {
   children?: ReactNode
 }
