+++ app/(marketing)/layout.tsx
@@ -3,11 +3,6 @@ import { Header } from "@/components/layout/header"
 
 import type { ReactNode } from "react"
 
-/**
- * Use the edge runtime
- */
-export const runtime = "edge"
-
 interface Props {
   children?: ReactNode
 }
