+++ app/api/ping/route.ts
@@ -8,10 +8,6 @@ interface ResponseData {
 }
 
 const gitSha = env.VERCEL_GIT_COMMIT_SHA ?? "local"
-
-/**
- * Use the edge runtime
- */
 export const runtime = "edge"
 
 /**
