+++ auth.config.ts
@@ -1,7 +1,9 @@
 import GitHub from "@auth/core/providers/github"
 import Google from "@auth/core/providers/google"
+import { DrizzleAdapter } from "@auth/drizzle-adapter"
+import { type NextAuthConfig } from "next-auth"
 
-import type { NextAuthConfig } from "next-auth"
+import { db, pgTable } from "@/lib/db"
 
 /**
  * This is the base-level config for NextAuth, but does not include a database
@@ -15,16 +17,26 @@ import type { NextAuthConfig } from "next-auth"
  */
 export default {
   /**
+   * https://authjs.dev/reference/adapter/drizzle
+   */
+  adapter: DrizzleAdapter(db, pgTable),
+
+  /**
    * NOTE: The email provider requires a database adapter. Because our database
    * adapter many not be edge-ready, we define that in the `./node.ts` file.
    */
   providers: [Google, GitHub],
+
   /**
    * Using JWTs for session tokens, so we can access the user's ID and email
    * from edge networks without requiring database access (which may not be
    * available in edge environments).
    */
   session: { strategy: "jwt" },
+
+  /**
+   *
+   */
   callbacks: {
     jwt({ token, user }) {
       // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
@@ -36,6 +48,10 @@ export default {
       return token
     }
   },
+
+  /**
+   *
+   */
   pages: {
     signIn: "/signin"
   }
