+++ package.json
@@ -39,7 +39,6 @@
     "next-themes": "^0.2.1",
     "nextjs-google-analytics": "^2.3.3",
     "nodemailer": "^6.9.4",
-    "postgres": "^3.3.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-email": "1.9.4",
@@ -73,6 +72,7 @@
     "next-secure-headers": "^2.2.0",
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.29",
+    "postgres": "^3.3.5",
     "prettier": "^3.0.3",
     "tailwindcss": "^3.3.3",
     "tailwindcss-animate": "^1.0.7",
